edge/requirements.txt
ultralytics
opencv-python
PyYAML
requests
supabase
python-dotenv
opencv-python